add companies permission

+++ app/controllers/api/v1/companies_controller.rb
@@ -1,18 +1,21 @@
 class Api::V1::CompaniesController < ApplicationController
-
+  before_action :authenticate_api_v1_user!, except: %i[index show]
   # before_action => コントローラの全てのアクションが実行される前に処理を行う
   before_action :set_post, only: [:show, :update, :destroy]
 
   def index
     companies = Company.order(created_at: :desc)
+    authorize companies
     render json: { companies: companies}
   end
 
   def show
+    authorize @company
     render json: { company: @company}
   end
 
   def create
+    authorize Company
     company = Company.new(company_params)
     if company.save
       render json: { status: 'success', company: company }
@@ -23,6 +26,7 @@ class Api::V1::CompaniesController < ApplicationController
   end
 
   def update
+    authorize @company
     if @company.update(company_params)
       render json: { status: 'success', company: @company }
     else
@@ -32,6 +36,7 @@ class Api::V1::CompaniesController < ApplicationController
   end
 
   def destroy
+    authorize @company
     if @company.destroy
       render json: { status: 'success', company: @company }
     else
